+++ cmd/contractcmd/deployBridge.go
@@ -4,8 +4,6 @@ package contractcmd
 
 import (
 	_ "embed"
-	"encoding/json"
-	"fmt"
 	"strings"
 
 	"github.com/ava-labs/avalanche-cli/pkg/cobrautils"
@@ -114,7 +112,7 @@ func CallDeployBridge(_ []string, flags DeployFlags) error {
 		nativeOption := "The native token " + tokenSymbol
 		erc20Option := "An ERC-20 token"
 		explainOption := "Explain the difference"
-		popularTokensInfo, err := getPopularTokensInfo(network, subnetOption)
+		popularTokensInfo, err := GetPopularTokensInfo(network, subnetOption)
 		if err != nil {
 			return err
 		}
@@ -182,34 +180,4 @@ func CallDeployBridge(_ []string, flags DeployFlags) error {
 		}
 	}
 	return nil
-}
-
-type PopularTokenInfo struct {
-	TokenName            string
-	TokenContractAddress string
-	BridgeHubAddress     string
-}
-
-//go:embed popularTokensInfo.json
-var popularTokensInfoByteSlice []byte
-
-var popularTokensInfo map[string][]PopularTokenInfo
-
-func (i PopularTokenInfo) Desc() string {
-	if i.TokenContractAddress == "" {
-		return i.TokenName
-	} else {
-		return fmt.Sprintf("%s | Token address %s | Hub bridge address %s", i.TokenName, i.TokenContractAddress, i.BridgeHubAddress)
-	}
-}
-
-func getPopularTokensInfo(network models.Network, subnetOption string) ([]PopularTokenInfo, error) {
-	if err := json.Unmarshal(popularTokensInfoByteSlice, &popularTokensInfo); err != nil {
-		return nil, fmt.Errorf("unabled to get popular tokens info from file: %w", err)
-	}
-	if network.Kind == models.Fuji && subnetOption == "C-Chain" {
-		return popularTokensInfo[models.Fuji.String()], nil
-	} else {
-		return nil, nil
-	}
 }
